fw/drivers/backlight: remove artificial scaling
All existing controller driver accept as input a % (0-100). While we
have slightly better resolution (e.g. 0-255 on AW2016) it was never used
anyway.

+++ src/fw/drivers/stubs/backlight.c
@@ -6,7 +6,7 @@
 void backlight_init(void) {
 }
 
-void backlight_set_brightness(uint16_t brightness) {
+void backlight_set_brightness(uint8_t brightness) {
 }
 
 void command_backlight_ctl(const char *arg) {
